+++ src/main/java/com/ym/hygg/huyagg/service/UserService.java
@@ -3,7 +3,7 @@ package com.ym.hygg.huyagg.service;
 import com.ym.hygg.huyagg.pojo.User;
 
 public interface UserService {
-    User getUserById(long id);
+    User getUserById(Integer id);
     int save(User user);
     User getUserByNameAndPassword(String username,String password);
 }
